Fix incorrectly moved BusSyncHandler interface.
